Convert Fit to Data Frame

Convert a parameter list into a peak data frame

param_list_to_peak_df(param_list, spec_names = NULL)

Arguments

param_list
list of fit parameters (or a list of such lists)

spec_names
character vector of spectrum names

Value

A data frame with the following columns:

peak

peak number

fit

fit cluster number, with all peaks in the same cluster having the same r2, scalar couplings, and m0

f_pvalue

optional, p-value determined from F-test during iterative fitting

omega0_ppm_1

chemical shift of singlet/doublet center in the first dimension (ppm)

omega0_ppm_2

chemical shift of singlet/doublet center in the second dimension (ppm)

sc_hz_1

optional, scalar coupling of doublet in first dimension (Hz)

sc_hz_2

optional, scalar coupling of doublet in second dimension (Hz)

r2_hz_1

R2 in first dimension (Hz)

r2_hz_2

R2 in first dimension (Hz)

...

m0 values for each spectrum

Details

This function takes the input or (if present) output parameters from a fit and converts them into a data frame. A parameter list must contain three lists:

start_list or fit_list

input or output values of the respective fit parameters

group_list

group numbers for the fit parameters

comb_list

coefficients for deriving fit parameters from a linear combination other auxiliary parameters

This function currently assumes the fit parameters were generated by fit_peak_iter, fit_peak_cluster, or peak_df_to_param_list. These functions use a particular convention for group_list and comb_list to represent either singlets or doublets in each dimension of a 2D spectrum.

This function can take either a single parameter list or a list of parameter lists. If the latter is given, then the results from all the parameter lists will be combined into a single table.